+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.0
+
+- Breaking refactor by replacing FlutterPlatformAlertOption with PlatformAlertOptions.
+- New customizations with IosAlertStyle and IosButtonStyle.
+- Updates documentation and examples.
+
 ## 0.4.0
 
 - Support Flutter 3.10.x.